refactor(state): split the disk_format module
