+++ zebra-state/src/service/finalized_state/disk_format/tests.rs
@@ -1,0 +1,3 @@
+//! Tests for the finalized disk format.
+
+mod prop;
