+++ zebra-state/src/service/finalized_state/tests.rs
@@ -1,2 +1,4 @@
+//! Finalized state tests.
+
 mod prop;
 mod vectors;
